refactor(lidar_apollo_instance_segmentation)!: fix namespace and directory structure (#7995)
* refactor: add autoware namespace prefix


* chore: update CODEOWNERS


* refactor: add `autoware` prefix


---------
