+++ launch/tier4_perception_launch/package.xml
@@ -22,6 +22,7 @@
   <exec_depend>autoware_euclidean_cluster</exec_depend>
   <exec_depend>autoware_ground_segmentation</exec_depend>
   <exec_depend>autoware_image_projection_based_fusion</exec_depend>
+  <exec_depend>autoware_lidar_apollo_instance_segmentation</exec_depend>
   <exec_depend>autoware_map_based_prediction</exec_depend>
   <exec_depend>autoware_multi_object_tracker</exec_depend>
   <exec_depend>autoware_object_merger</exec_depend>
@@ -37,7 +38,6 @@
   <exec_depend>autoware_tracking_object_merger</exec_depend>
   <exec_depend>elevation_map_loader</exec_depend>
   <exec_depend>image_transport_decompressor</exec_depend>
-  <exec_depend>lidar_apollo_instance_segmentation</exec_depend>
   <exec_depend>occupancy_grid_map_outlier_filter</exec_depend>
   <exec_depend>pointcloud_to_laserscan</exec_depend>
   <exec_depend>probabilistic_occupancy_grid_map</exec_depend>
