+++ perception/autoware_lidar_apollo_instance_segmentation/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>lidar_apollo_instance_segmentation</name>
+  <name>autoware_lidar_apollo_instance_segmentation</name>
   <version>0.1.0</version>
-  <description>lidar_apollo_instance_segmentation</description>
+  <description>autoware_lidar_apollo_instance_segmentation</description>
   <maintainer email="yoshi.ri@tier4.jp">Yoshi Ri</maintainer>
   <maintainer email="yukihiro.saito@tier4.jp">Yukihiro Saito</maintainer>
   <license>Apache License 2.0</license>
